refactor: 대기열 리펙토링
