+++ application/gateway-service/src/main/java/com/fortickets/gatewayservice/kafka/KafkaConfig.java
@@ -1,5 +1,6 @@
-package com.fortickets.gatewayservice.newkafka;
+package com.fortickets.gatewayservice.kafka;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.HashMap;
 import java.util.Map;
 import org.apache.kafka.clients.admin.NewTopic;
@@ -63,4 +64,7 @@ public class KafkaConfig {
         configs.put("bootstrap.servers", bootstrapServers);
         return new KafkaAdmin(configs);
     }
+
+    @Bean
+    public ObjectMapper objectMapper() { return new ObjectMapper(); }
 }
