+++ application/gateway-service/src/main/java/com/fortickets/gatewayservice/kafka/KafkaFilter.java
@@ -1,6 +1,5 @@
-package com.fortickets.gatewayservice.newkafka;
+package com.fortickets.gatewayservice.kafka;
 
-import com.fortickets.gatewayservice.newkafka.KafkaFilter.Config;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
