+++ application/gateway-service/src/main/resources/application.yml
@@ -11,7 +11,6 @@ spring:
   cloud:
     gateway:
       default-filters:
-#        - name: TicketFilter
         - name: KafkaFilter
       routes:
         - id: user-service
